+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K828ENWM7K0ZW6X9XKH9RG93"
+        "01K82WV7KKNVS9Q61G57MJDMT3"
     )
     r"""The id of the resource"""