+++ packages/orq-rc/RELEASES.md
@@ -3449,3 +3449,13 @@ Based on:
 - [python v3.14.0-rc.40] packages/orq-rc
 ### Releases
 - [PyPI v3.14.0-rc.40] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.40 - packages/orq-rc
+
+## 2025-10-21 08:07:31
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.637.3 (2.727.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.41] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.41] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.41 - packages/orq-rc